fix: align mosca index float types and add neo4j dependency error handling

+++ backend/app/api/v1/endpoints/mosca.py
@@ -5,11 +5,11 @@ router = APIRouter(prefix="/mosca", tags=["mosca"])
 
 @router.get("/calculate")
 def calculate_mosca_index(
-    x: int = Query(
+    x: float = Query(
         ..., description="Security shelf-life (years data must remain secure)"
     ),
-    y: int = Query(..., description="Migration time (years required to transition)"),
-    z: int = Query(..., description="Threat horizon (years until Q-Day)"),
+    y: float = Query(..., description="Migration time (years required to transition)"),
+    z: float = Query(..., description="Threat horizon (years until Q-Day)"),
 ):
     """
     Calculate the Mosca Readiness Index given x, y, and z parameters.
